fix：修改了一些问题

+++ edu-admin/backend/src/main/java/com/scu/eduadmin/vo/GradeVO.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public class GradeVO {
     private Long selectionId;
     private String courseName;
-    private Integer credits;
+    private BigDecimal credits;
     private String semester;
     private BigDecimal score;
 }
